Rename and move files to disconnect CMake and compilation database request
